refactor: extract fs core module for layered architecture
Create a dedicated src/fs/ module as Layer 1 of the file system architecture.
This module contains all core file system domain logic with no dependencies on
HTTP or AI interfaces.

Files moved:
- src/models/files.rs -> src/fs/models.rs
- src/queries/files.rs -> src/fs/queries.rs
- src/services/files.rs -> src/fs/services.rs
- src/services/storage.rs -> src/fs/storage.rs
- src/parsers/{links,tags}.rs -> src/fs/parsers/
- src/workers/{link,tag}_indexer.rs -> src/fs/workers/
- src/utils/{document_metadata,frontmatter,yaml_frontmatter}.rs -> src/fs/utils/

Architecture:
- Layer 1 (fs): Core file system logic (models, queries, services, storage)
- Layer 2 (handlers/tools): REST and AI interfaces consume fs
- Layer 3 (agents): AI agents use tools

Benefits:
- Single source of truth for file system logic
- Interface independence (REST and AI can evolve independently)
- Better testability without HTTP or AI concerns
- Clear dependency hierarchy
