+++ backend/src/fs/utils/document_metadata.rs
