+++ backend/src/models/mod.rs
@@ -1,7 +1,6 @@
 pub mod agent_session;
 pub mod ai_models;
 pub mod chat;
-pub mod files;
 pub mod invitations;
 pub mod permissions;
 pub mod requests;
@@ -10,3 +9,8 @@ pub mod sse;
 pub mod users;
 pub mod workspace_members;
 pub mod workspaces;
+
+// Re-export file models from fs module for backward compatibility
+pub mod files {
+    pub use crate::fs::models::{File, FileType, NewFile, UpdateFileContent};
+}
