+++ backend/src/parsers/mod.rs
@@ -2,9 +2,9 @@
 //!
 //! This module provides parsers for extracting links and tags from markdown content,
 //! following Obsidian's conventions.
+//!
+//! Note: The actual implementations are now in the `fs::parsers` module.
+//! This module re-exports them for backward compatibility.
 
-mod links;
-mod tags;
-
-pub use links::{extract_links, extract_links_with_display};
-pub use tags::extract_tags;
+// Re-export parsers from fs module
+pub use crate::fs::parsers::{extract_links, extract_links_with_display, extract_tags, Wikilink};
