+++ backend/src/queries/mod.rs
@@ -1,10 +1,14 @@
 pub mod agent_sessions;
 pub mod ai_models;
 pub mod chat;
-pub mod files;
 pub mod invitations;
 pub mod roles;
 pub mod sessions;
 pub mod users;
 pub mod workspaces;
 pub mod workspace_members;
+
+// Re-export file queries from fs module for backward compatibility
+pub mod files {
+    pub use crate::fs::queries::*;
+}
