+++ backend/src/services/mod.rs
@@ -1,7 +1,6 @@
 pub mod agent_sessions;
 pub mod chat;
 pub mod cookies;
-pub mod files;
 pub mod invitations;
 pub mod jwt;
 pub mod refresh_tokens;
@@ -10,4 +9,12 @@ pub mod roles;
 pub mod workspaces;
 pub mod workspace_members;
 pub mod sessions;
-pub mod storage;
+
+// Re-export file services from fs module for backward compatibility
+pub mod files {
+    pub use crate::fs::services::*;
+}
+
+pub mod storage {
+    pub use crate::fs::storage::FileStorageService;
+}
