+++ backend/src/workers/mod.rs
@@ -1,9 +1,8 @@
 pub mod revoked_token_cleanup;
 pub mod archive_cleanup;
-pub mod tag_indexer;
-pub mod link_indexer;
 
 pub use revoked_token_cleanup::revoked_token_cleanup_worker;
 pub use archive_cleanup::archive_cleanup_worker;
-pub use tag_indexer::tag_indexer_worker;
-pub use link_indexer::link_indexer_worker;
+
+// Re-export file system workers from fs module
+pub use crate::fs::workers::{tag_indexer_worker, link_indexer_worker};
